Moved start build endpoints to build.go
